requirements.txt
firebase-admin==6.2.0
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
requests